# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/plants": {
      "get": {
        "operationId": "EnergyController_listPlants",
        "parameters": [],
        "responses": {
          "200": {
            "description": "电站列表"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "查询当前用户名下电站",
        "tags": [
          "energy"
        ]
      },
      "post": {
        "operationId": "EnergyController_createPlant",
        "parameters": [],
        "responses": {
          "201": {
            "description": "创建成功"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "创建电站",
        "tags": [
          "energy"
        ]
      }
    },
    "/plants/{id}": {
      "get": {
        "operationId": "EnergyController_getPlant",
        "parameters": [],
        "responses": {
          "200": {
            "description": "电站信息"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "按 id 查询电站",
        "tags": [
          "energy"
        ]
      },
      "patch": {
        "operationId": "EnergyController_updatePlant",
        "parameters": [],
        "responses": {
          "200": {
            "description": "更新成功"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "更新电站",
        "tags": [
          "energy"
        ]
      },
      "delete": {
        "operationId": "EnergyController_removePlant",
        "parameters": [],
        "responses": {
          "204": {
            "description": "删除成功"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "删除电站",
        "tags": [
          "energy"
        ]
      }
    },
    "/devices": {
      "get": {
        "operationId": "EnergyController_listUserDevices",
        "parameters": [],
        "responses": {
          "200": {
            "description": "设备列表"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "查询当前用户名下设备",
        "tags": [
          "energy"
        ]
      },
      "post": {
        "operationId": "EnergyController_createDevice",
        "parameters": [],
        "responses": {
          "201": {
            "description": "创建成功"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "创建设备（挂靠在电站下）",
        "tags": [
          "energy"
        ]
      }
    },
    "/plants/{plantId}/devices": {
      "get": {
        "operationId": "EnergyController_listDevices",
        "parameters": [],
        "responses": {
          "200": {
            "description": "设备列表"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "查询电站下设备列表",
        "tags": [
          "energy"
        ]
      }
    },
    "/devices/{id}": {
      "get": {
        "operationId": "EnergyController_getDevice",
        "parameters": [],
        "responses": {
          "200": {
            "description": "设备信息"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "按 id 查询设备",
        "tags": [
          "energy"
        ]
      },
      "patch": {
        "operationId": "EnergyController_updateDevice",
        "parameters": [],
        "responses": {
          "200": {
            "description": "更新成功"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "更新设备",
        "tags": [
          "energy"
        ]
      },
      "delete": {
        "operationId": "EnergyController_removeDevice",
        "parameters": [],
        "responses": {
          "204": {
            "description": "删除成功"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "删除设备",
        "tags": [
          "energy"
        ]
      }
    },
    "/tariffs": {
      "get": {
        "operationId": "EnergyController_listTariffs",
        "parameters": [],
        "responses": {
          "200": {
            "description": "费率列表"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "查询全部费率（按生效时间倒序）",
        "tags": [
          "energy"
        ]
      },
      "post": {
        "operationId": "EnergyController_createTariff",
        "parameters": [],
        "responses": {
          "201": {
            "description": "创建成功"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "创建费率",
        "tags": [
          "energy"
        ]
      }
    },
    "/tariffs/{id}": {
      "delete": {
        "operationId": "EnergyController_removeTariff",
        "parameters": [],
        "responses": {
          "204": {
            "description": "删除成功"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "删除费率",
        "tags": [
          "energy"
        ]
      }
    },
    "/energy-records": {
      "post": {
        "operationId": "EnergyController_recordEnergy",
        "parameters": [],
        "responses": {
          "201": {
            "description": "写入成功"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "写入发电记录（自动建分区）",
        "tags": [
          "energy"
        ]
      }
    },
    "/plants/{plantId}/energy-records": {
      "get": {
        "operationId": "EnergyController_listEnergyRecords",
        "parameters": [],
        "responses": {
          "200": {
            "description": "发电记录列表"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "查询电站发电记录（区间）",
        "tags": [
          "energy"
        ]
      }
    },
    "/plants/{plantId}/energy/daily": {
      "get": {
        "operationId": "EnergyController_dailyAggregate",
        "parameters": [],
        "responses": {
          "200": {
            "description": "按天聚合结果"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "电站日发电聚合",
        "tags": [
          "energy"
        ]
      }
    },
    "/plants/{plantId}/energy/total": {
      "get": {
        "operationId": "EnergyController_totalAggregate",
        "parameters": [],
        "responses": {
          "200": {
            "description": "总发电量汇总"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "电站区间总发电量",
        "tags": [
          "energy"
        ]
      }
    },
...
    "/anomaly-rules": {
      "get": {
        "operationId": "AnomalyController_listRules",
        "parameters": [],
        "responses": {
          "200": {
            "description": "规则列表"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "查询检测规则",
        "tags": [
          "anomaly"
        ]
      },
      "post": {
        "operationId": "AnomalyController_createRule",
        "parameters": [],
        "responses": {
          "201": {
            "description": "创建成功"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "创建检测规则",
        "tags": [
          "anomaly"
        ]
      }
    },
    "/anomaly-rules/{id}": {
      "patch": {
        "operationId": "AnomalyController_updateRule",
        "parameters": [],
        "responses": {
          "200": {
            "description": "更新成功"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "更新检测规则（版本 +1）",
        "tags": [
          "anomaly"
        ]
      },
      "delete": {
        "operationId": "AnomalyController_removeRule",
        "parameters": [],
        "responses": {
          "204": {
            "description": "删除成功"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "删除检测规则",
        "tags": [
          "anomaly"
        ]
      }
    },
    "/anomaly/detect": {
      "post": {
        "operationId": "AnomalyController_detect",
        "parameters": [],
        "responses": {
          "201": {
            "description": "命中的告警（已落库）"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "触发即时异常检测",
        "tags": [
          "anomaly"
        ]
      }
    },
    "/anomalies": {
      "get": {
        "operationId": "AnomalyController_listUserEvents",
        "parameters": [],
        "responses": {
          "200": {
            "description": "异常事件列表"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "查询当前用户异常事件",
        "tags": [
          "anomaly"
        ]
      }
    },
    "/plants/{plantId}/anomaly-events": {
      "get": {
        "operationId": "AnomalyController_listEvents",
        "parameters": [],
        "responses": {
          "200": {
            "description": "异常事件列表"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "查询电站异常事件",
        "tags": [
          "anomaly"
        ]
      }
    }
  },
  "info": {
    "title": "Helio API",
    "description": "Helio 太阳能能源监控平台 API",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
...
  }
}